requirements.txt
selene
pytest~=8.3.5
selenium~=4.31.0
allure-pytest
allure-python-commons~=2.14.0
python-dotenv~=1.1.0